Bug warning

+++ project/plugins/acVinLibPlugin/lib/csv/CsvFile.class.php
@@ -15,6 +15,7 @@ class CsvFile
   }
 
   public function __construct($file = null, $options = array()) {
+    $this->isFileTmp = false;
     $this->options = $options;
     if (!isset($this->options["ignore_first_if_comment"])) {
       $this->options["ignore_first_if_comment"] = true;
@@ -76,7 +77,7 @@ class CsvFile
     }
     fclose($handler);
 
-    if($this->isFileTmp) {
+    if($this->isFileTmp && $this->file) {
         unlink($this->file);
     }
     return $this->csvdata;
@@ -90,14 +91,14 @@ class CsvFile
     }
     $ret = exec('file -i '.$file);
     $charset = substr($ret, strpos($ret,'charset='));
-    if(isset($fileTmp)) {
+    if(isset($fileTmp) && $fileTmp) {
         unlink($fileTmp);
     }
     return str_replace('charset=','',$charset);
   }
 
     public function __destruct() {
-        if($this->isFileTmp) {
+        if($this->isFileTmp && $this->file) {
             unlink($this->file);
         }
     }
